requirements.txt
asgiref==3.11.1
Django==6.0.3
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.17.0
djangorestframework-simplejwt==5.4.0
pillow==12.1.1
sqlparse==0.5.5
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
psycopg2-binary==2.9.10
whitenoise==6.9.0
requests
django-redis==6.0.0
